+++ src/SimpleJWT/Crypt/AESGCM.php
@@ -2,7 +2,7 @@
 /*
  * SimpleJWT
  *
- * Copyright (C) Kelvin Mo 2020
+ * Copyright (C) Kelvin Mo 2020-2021
  *
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions
@@ -107,7 +107,7 @@ class AESGCM extends Algorithm implements EncryptionAlgorithm {
         $tag = Util::base64url_decode($tag);
         if (strlen($tag) != 16) throw new CryptException('Incorrect authentication tag length');
         
-        $plaintext = openssl_decrypt($ciphertext, $params['cipher'], $cek, OPENSSL_RAW_DATA, $iv, $tag, $additional);
+        $plaintext = openssl_decrypt(Util::base64url_decode($ciphertext), $params['cipher'], $cek, OPENSSL_RAW_DATA, $iv, $tag, $additional);
         if ($plaintext === false) throw new CryptException('Authentication tag does not match');
 
         return $plaintext;