+++ CHANGELOG.md
@@ -2,7 +2,11 @@
 
 All notable changes to this project will be documented in this file.
 
-## Unreleased
+## Version 0.5.1
+
+- Added: Support for PHP 8 (#35)
+
+## Version 0.5.0
 
 - Added: Support for AES GCM family of algorithms
 - Added: Support for Elliptic Curve Diffie-Hellman key derivation
@@ -12,7 +16,10 @@ All notable changes to this project will be documented in this file.
 - Changed: SimpleJWT\JWT::deserialise() no longer supports $format parameter
   (format is automatically detected)
 - Changed: Return value of SimpleJWT\JWT::deserialise() changed
+- Changed: SimpleJWT\JWE::decrypt() no longer supports $format parameter
+  (format is automatically detected)
 - Removed: SimpleJWT\Keys\Key::getSignature()
+- Fixed: Autoload issue in jwkstool (#31)
 
 ## Version 0.4.2